feat: adds support for firebase auth emulator
