+++ src/utils.ts
@@ -21,3 +21,18 @@ export function fetchPublicKeys(
       });
   });
 }
+
+export function decodeJwtHeader(jwt: string) {
+  const [base64] = jwt.split('.');
+  const buffer = Buffer.from(base64, 'base64');
+  return buffer.toString();
+}
+
+export function getBearerToken(auth?: string) {
+  if (auth) {
+    const [scheme, credentials] = auth.split(' ');
+    if (scheme.toLocaleLowerCase() === 'bearer') {
+      return credentials;
+    }
+  }
+}
